+++ server/libs/modules/components/zoho/zoho-crm/src/test/java/com/bytechef/component/zoho/crm/ZohoCrmComponentHandlerTest.java
@@ -14,21 +14,18 @@
  * limitations under the License.
  */
 
-package com.bytechef.component.zoho.constant;
+package com.bytechef.component.zoho.crm;
+
+import com.bytechef.test.jsonasssert.JsonFileAssert;
+import org.junit.jupiter.api.Test;
 
 /**
  * @author Luka Ljubić
  */
-public class ZohoCrmConstants {
+class ZohoCrmComponentHandlerTest {
 
-    public static final String EMAIL = "email";
-    public static final String FIRST_NAME = "first_name";
-    public static final String LAST_NAME = "last_name";
-    public static final String PROFILE = "profile";
-    public static final String REGION = "region";
-    public static final String ROLE = "role";
-    public static final String TYPE = "type";
-
-    private ZohoCrmConstants() {
+    @Test
+    void testGetComponentDefinition() {
+        JsonFileAssert.assertEquals("definition/zoho-crm_v1.json", new ZohoCrmComponentHandler().getDefinition());
     }
 }
