+++ server/libs/modules/components/zoho/zoho-crm/src/test/java/com/bytechef/component/zoho/crm/action/ZohoCrmGetOrganizationActionTest.java
@@ -14,13 +14,8 @@
  * limitations under the License.
  */
 
-package com.bytechef.component.zoho.action;
+package com.bytechef.component.zoho.crm.action;
 
-import static com.bytechef.component.zoho.constant.ZohoCrmConstants.EMAIL;
-import static com.bytechef.component.zoho.constant.ZohoCrmConstants.FIRST_NAME;
-import static com.bytechef.component.zoho.constant.ZohoCrmConstants.LAST_NAME;
-import static com.bytechef.component.zoho.constant.ZohoCrmConstants.PROFILE;
-import static com.bytechef.component.zoho.constant.ZohoCrmConstants.ROLE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
@@ -30,33 +25,23 @@ import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.Context.Http;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.definition.TypeReference;
-import com.bytechef.component.test.definition.MockParametersFactory;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
 
 /**
  * @author Luka Ljubić
  * @author Monika Kušter
  */
-class ZohoCrmAddUserActionTest {
+class ZohoCrmGetOrganizationActionTest {
 
-    private final ArgumentCaptor<Http.Body> bodyArgumentCaptor = ArgumentCaptor.forClass(Http.Body.class);
     private final ActionContext mockedActionContext = mock(ActionContext.class);
     private final Http.Executor mockedExecutor = mock(Http.Executor.class);
     private final Object mockedObject = mock(Object.class);
-    private final Parameters mockedParameters = MockParametersFactory.create(
-        Map.of(ROLE, "user_role", FIRST_NAME, "first_name", EMAIL, "email",
-            PROFILE, "user_profile", LAST_NAME, "last_name"));
+    private final Parameters mockedParameters = mock(Parameters.class);
     private final Http.Response mockedResponse = mock(Http.Response.class);
 
     @Test
     void testPerform() {
         when(mockedActionContext.http(any()))
-            .thenReturn(mockedExecutor);
-        when(mockedExecutor.body(bodyArgumentCaptor.capture()))
             .thenReturn(mockedExecutor);
         when(mockedExecutor.configuration(any()))
             .thenReturn(mockedExecutor);
@@ -65,20 +50,8 @@ class ZohoCrmAddUserActionTest {
         when(mockedResponse.getBody(any(TypeReference.class)))
             .thenReturn(mockedObject);
 
-        Object result = ZohoCrmAddUserAction.perform(mockedParameters, mockedParameters, mockedActionContext);
+        Object result = ZohoCrmGetOrganizationAction.perform(mockedParameters, mockedParameters, mockedActionContext);
 
         assertEquals(mockedObject, result);
-
-        Http.Body body = bodyArgumentCaptor.getValue();
-
-        Map<String, Object> userMap = new HashMap<>();
-
-        userMap.put(ROLE, "user_role");
-        userMap.put(FIRST_NAME, "first_name");
-        userMap.put(EMAIL, "email");
-        userMap.put(PROFILE, "user_profile");
-        userMap.put(LAST_NAME, "last_name");
-
-        assertEquals(Map.of("users", List.of(userMap)), body.getContent());
     }
 }
