+++ server/libs/modules/components/zoho/zoho-crm/src/test/java/com/bytechef/component/zoho/crm/util/ZohoCrmUtilTest.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.bytechef.component.zoho.action;
+package com.bytechef.component.zoho.crm.util;
 
-import static com.bytechef.component.zoho.constant.ZohoCrmConstants.TYPE;
+import static com.bytechef.component.definition.ComponentDsl.option;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
@@ -24,47 +24,52 @@ import static org.mockito.Mockito.when;
 
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.Context.Http;
+import com.bytechef.component.definition.Option;
 import com.bytechef.component.definition.Parameters;
 import com.bytechef.component.definition.TypeReference;
-import com.bytechef.component.test.definition.MockParametersFactory;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
 
 /**
  * @author Luka Ljubić
- * @author Monika Kušter
  */
-class ZohoCrmListUsersActionTest {
+class ZohoCrmUtilTest {
 
+    private final List<Option<String>> expectedOptions = List.of(option("option", "123"));
     private final ActionContext mockedActionContext = mock(ActionContext.class);
     private final Http.Executor mockedExecutor = mock(Http.Executor.class);
-    private final Object mockedObject = mock(Object.class);
-    private final Parameters mockedParameters = MockParametersFactory.create(Map.of(TYPE, "type"));
+    private final Parameters mockedParameters = mock(Parameters.class);
     private final Http.Response mockedResponse = mock(Http.Response.class);
-    private final ArgumentCaptor<Object[]> queryArgumentCaptor = ArgumentCaptor.forClass(Object[].class);
 
-    @Test
-    void testPerform() {
+    @BeforeEach
+    void beforeEach() {
         when(mockedActionContext.http(any()))
-            .thenReturn(mockedExecutor);
-        when(mockedExecutor.queryParameters(queryArgumentCaptor.capture()))
             .thenReturn(mockedExecutor);
         when(mockedExecutor.configuration(any()))
             .thenReturn(mockedExecutor);
         when(mockedExecutor.execute())
             .thenReturn(mockedResponse);
+    }
+
+    @Test
+    void testGetProfileOptions() {
         when(mockedResponse.getBody(any(TypeReference.class)))
-            .thenReturn(mockedObject);
+            .thenReturn(Map.of("profiles", List.of(Map.of("name", "option", "id", "123"))));
 
-        Object result = ZohoCrmListUsersAction.perform(mockedParameters, mockedParameters, mockedActionContext);
+        assertEquals(
+            expectedOptions,
+            ZohoCrmUtils.getProfileOptions(mockedParameters, mockedParameters, Map.of(), "", mockedActionContext));
+    }
 
-        assertEquals(mockedObject, result);
+    @Test
+    void testGetRoleOptions() {
+        when(mockedResponse.getBody(any(TypeReference.class)))
+            .thenReturn(Map.of("roles", List.of(Map.of("name", "option", "id", "123"))));
 
-        Object[] query = queryArgumentCaptor.getValue();
-
-        assertEquals(List.of(TYPE, "type"), Arrays.asList(query));
+        assertEquals(
+            expectedOptions,
+            ZohoCrmUtils.getRoleOptions(mockedParameters, mockedParameters, Map.of(), "", mockedActionContext));
     }
 }
